net: banip: Add support for nft based tables.
This is very rough.  Limitations:
  - No attempt is made to handle the auto-ban or email notifications.
  - It uses a LOT more memory. Realistically 512MB is minimum.
  - Set counters are disabled.  They require a kernel patch that
    is unlikely to be accepted due to nft bug.
  - Requires ban_backupdir in persistent storage to really work properly.
  - Ban rules run AFTER fw4.  I think this is a feature but if you like
    watching bozo's hammer on your firewall you may think different.

It will load sets and work with existing luci interface.

There is exactly one new feature:  ban_country_whitelist <bool>. If
this is set to true/1 then the country lists will be interpreted as
allowing ONLY the specified countries.  If it is false/0 or misssing
then prior behavior of blocking specified countries will be used.